+++ demo.js
@@ -16,5 +16,6 @@ console.log(`http://localhost:${port}/read-only.html`);
 console.log(`http://localhost:${port}/row-filtering-example.html`);
 console.log(`http://localhost:${port}/bitmask-example.html`);
 console.log(`http://localhost:${port}/set-annotations.html`);
+console.log(`http://localhost:${port}/class-focus.html`);
 console.log(`http://localhost:${port}/live_demo.html`);
 console.log(`http://localhost:${port}/offset-container.html`);